Add state machine for imports
Due to limiting lookahead functionality of imports at the moment
we are not allowing full status transition functionality for imports
and only using a limited subset of allowed transitions at the moment.

We still need to have None to In Progress transition because
validate_status still executes full SM before saving values to the
database.